+++ src/test/java/nextstep/ladder/domain/LadderConfigTest.java
@@ -1,6 +1,5 @@
-package nextstep.ladder.domain.edge;
+package nextstep.ladder.domain;
 
-import nextstep.ladder.domain.LadderConfig;
 import nextstep.ladder.domain.prize.LadderPrize;
 import nextstep.ladder.domain.prize.LadderPrizes;
 import nextstep.ladder.domain.user.LadderUser;
