+++ src/test/java/nextstep/ladder/domain/line/LinesTest.java
@@ -27,7 +27,8 @@ class LinesTest {
         new LadderUser("honux", 1),
         new LadderUser("crong", 2),
         new LadderUser("jk", 3)
-    ));;
+    ));
+    ;
 
     @DisplayName("Lines 인스턴스 만들기")
     @Test
