+++ src/test/java/nextstep/ladder/domain/line/PointTest.java
@@ -3,7 +3,7 @@ package nextstep.ladder.domain.line;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 class PointTest {
