+++ src/test/java/nextstep/ladder/domain/prize/LadderPrizeTest.java
@@ -1,10 +1,9 @@
-package nextstep.ladder.domain.edge;
+package nextstep.ladder.domain.prize;
 
-import nextstep.ladder.domain.prize.LadderPrize;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 class LadderPrizeTest {
