+++ src/test/java/nextstep/ladder/domain/user/PositionTest.java
@@ -1,8 +1,7 @@
-package nextstep.ladder.domain.edge;
+package nextstep.ladder.domain.user;
 
 import nextstep.ladder.domain.line.Line;
 import nextstep.ladder.domain.line.Point;
-import nextstep.ladder.domain.user.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -10,7 +9,8 @@ import java.util.List;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 class PositionTest {
 
@@ -22,7 +22,7 @@ class PositionTest {
 
     @DisplayName("Position 인스턴스 생성 - 음수 값은 들어갈 수 없음")
     @Test
-    public void testConstructor_throwException () {
+    public void testConstructor_throwException() {
         assertThatThrownBy(() -> new Position(-1))
             .isInstanceOf(IllegalArgumentException.class)
             .hasMessageContaining("위치 값은 0 이하가 될 수 없습니다.");
