+++ src/test/java/nextstep/ladder/factory/PointFactoryTest.java
@@ -1,6 +1,6 @@
-package nextstep.ladder.domain.line;
+package nextstep.ladder.factory;
 
-import nextstep.ladder.factory.PointFactory;
+import nextstep.ladder.domain.line.Point;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
